fix: nix install and caching
The previous commit quickly replaced the magic nix cache with the
flakehub cache, but I didn't realize that was a paid product. This
switches to the cache-nix-action maintained by the nix community.

This action has a dependency on a specific action to install nix. This
seems like it'll still support our use-case. The DeterminateSystems
installer was only used because it was the only one I found via google
at the time.